Refactor Schema class to pubsub_experimental
The whole proto is marked as "EXPERIMENTAL", all caps, I think we should
make the classes experimental too.
