+++ google/cloud/pubsub/experimental/schema_test.cc
@@ -12,13 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "google/cloud/pubsub/schema.h"
+#include "google/cloud/pubsub/experimental/schema.h"
 #include <gmock/gmock.h>
 #include <sstream>
 
 namespace google {
 namespace cloud {
-namespace pubsub {
+namespace pubsub_experimental {
 inline namespace GOOGLE_CLOUD_CPP_PUBSUB_NS {
 namespace {
 
@@ -56,6 +56,6 @@ TEST(Schema, OutputStream) {
 
 }  // namespace
 }  // namespace GOOGLE_CLOUD_CPP_PUBSUB_NS
-}  // namespace pubsub
+}  // namespace pubsub_experimental
 }  // namespace cloud
 }  // namespace google
