+++ google/cloud/pubsub/google_cloud_cpp_pubsub.bzl
@@ -21,6 +21,7 @@ google_cloud_cpp_pubsub_hdrs = [
     "application_callback.h",
     "backoff_policy.h",
     "connection_options.h",
+    "experimental/schema.h",
     "internal/batch_sink.h",
     "internal/batching_publisher_connection.h",
     "internal/create_channel.h",
@@ -53,7 +54,6 @@ google_cloud_cpp_pubsub_hdrs = [
     "publisher_connection.h",
     "publisher_options.h",
     "retry_policy.h",
-    "schema.h",
     "snapshot.h",
     "snapshot_builder.h",
     "subscriber.h",
@@ -74,6 +74,7 @@ google_cloud_cpp_pubsub_hdrs = [
 google_cloud_cpp_pubsub_srcs = [
     "ack_handler.cc",
     "connection_options.cc",
+    "experimental/schema.cc",
     "internal/batching_publisher_connection.cc",
     "internal/create_channel.cc",
     "internal/default_batch_sink.cc",
@@ -102,7 +103,6 @@ google_cloud_cpp_pubsub_srcs = [
     "publisher.cc",
     "publisher_connection.cc",
     "publisher_options.cc",
-    "schema.cc",
     "snapshot.cc",
     "snapshot_builder.cc",
     "subscriber_connection.cc",
